hollowforge/backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
aiosqlite>=0.20
httpx>=0.28
pillow>=11.0
python-multipart>=0.0.18
